Add timestamps

+++ include/display_record.hpp
@@ -1,11 +1,12 @@
 #pragma once
+#include <cstdint>
 #include <chrono>
 #include "can_message.hpp"
 
 struct DisplayRecord {
     CANMessage msg;
     std::chrono::steady_clock::time_point timestamp;
-    unsigned long timeDelta;
+    uint32_t timeDelta;
     uint8_t highlightBitmap;
 };
